cfsync starting to form
Instead of listing all the directories and sorting them like we were
doing in cdnalizerd, we're going to do a depth first sort and process.

That way we won't need to keep the already processed sub dirs in memory,
and we'll still be able to go through the sorted lists, comparing with
the remote version.

This commit should just print out local and remote file names, but it
hasn't been tested yet.